requirement.txt
behave==1.2.6
selenium==4.10.0
webdriver-manager==3.8.6
python-dotenv==1.0.0
allure-behave==2.13.2
requests==2.31.0
pytest==7.4.0
pytest-html==4.0.2